refactor(behavior_path_planner): separate safety check from util (#3331)
* refactor(behavior_path_planner): separate safety check from util


* revert non related getPredictedPathFromObj


---------
