+++ planning/behavior_path_planner/include/behavior_path_planner/utilities.hpp
@@ -73,11 +73,7 @@ using geometry_msgs::msg::Vector3;
 using route_handler::RouteHandler;
 using tier4_autoware_utils::LinearRing2d;
 using tier4_autoware_utils::LineString2d;
-using tier4_autoware_utils::Point2d;
 using tier4_autoware_utils::Polygon2d;
-namespace bg = boost::geometry;
-using geometry_msgs::msg::Pose;
-using marker_utils::CollisionCheckDebug;
 using vehicle_info_util::VehicleInfo;
 
 struct FrenetPoint
@@ -86,22 +82,7 @@ struct FrenetPoint
   double distance{0.0};  // lateral
 };
 
-struct ProjectedDistancePoint
-{
-  Point2d projected_point;
-  double distance{0.0};
-};
-
-template <typename Pythagoras = bg::strategy::distance::pythagoras<>>
-ProjectedDistancePoint pointToSegment(
-  const Point2d & reference_point, const Point2d & point_from_ego,
-  const Point2d & point_from_object);
-
-void getProjectedDistancePointFromPolygons(
-  const Polygon2d & ego_polygon, const Polygon2d & object_polygon, Pose & point_on_ego,
-  Pose & point_on_object);
 // data conversions
-
 PredictedPath convertToPredictedPath(
   const PathWithLaneId & path, const Twist & vehicle_twist, const Pose & pose,
   const size_t nearest_seg_idx, const double duration, const double resolution,
@@ -352,55 +333,8 @@ lanelet::ConstLanelets calcLaneAroundPose(
 std::vector<PredictedPath> getPredictedPathFromObj(
   const PredictedObject & obj, const bool & is_use_all_predicted_path);
 
-Pose projectCurrentPoseToTarget(const Pose & desired_object, const Pose & target_object);
-
 boost::optional<std::pair<Pose, Polygon2d>> getEgoExpectedPoseAndConvertToPolygon(
   const PredictedPath & pred_path, const double current_time, const VehicleInfo & ego_info);
-
-boost::optional<std::pair<Pose, Polygon2d>> getObjectExpectedPoseAndConvertToPolygon(
-  const PredictedPath & pred_path, const double current_time, const PredictedObject & object);
-
-bool isObjectFront(const Pose & ego_pose, const Pose & obj_pose);
-
-bool isObjectFront(const Pose & projected_ego_pose);
-
-double stoppingDistance(const double & vehicle_velocity, const double & vehicle_accel);
-
-double frontVehicleStopDistance(
-  const double & front_vehicle_velocity, const double & front_vehicle_accel,
-  const double & distance_to_collision);
-
-double rearVehicleStopDistance(
-  const double & rear_vehicle_velocity, const double & rear_vehicle_accel,
-  const double & rear_vehicle_reaction_time, const double & rear_vehicle_safety_time_margin);
-
-bool isLongitudinalDistanceEnough(
-  const double & rear_vehicle_stop_threshold, const double & front_vehicle_stop_threshold);
-
-bool hasEnoughDistance(
-  const Pose & expected_ego_pose, const Twist & ego_current_twist,
-  const Pose & expected_object_pose, const Twist & object_current_twist,
-  const BehaviorPathPlannerParameters & param, const double front_decel, const double rear_decel,
-  CollisionCheckDebug & debug);
-
-bool isLateralDistanceEnough(
-  const double & relative_lateral_distance, const double & lateral_distance_threshold);
-
-bool isSafeInLaneletCollisionCheck(
-  const std::vector<std::pair<Pose, tier4_autoware_utils::Polygon2d>> & interpolated_ego,
-  const Twist & ego_current_twist, const std::vector<double> & check_duration,
-  const double prepare_duration, const PredictedObject & target_object,
-  const PredictedPath & target_object_path, const BehaviorPathPlannerParameters & common_parameters,
-  const double prepare_phase_ignore_target_speed_thresh, const double front_decel,
-  const double rear_decel, Pose & ego_pose_before_collision, CollisionCheckDebug & debug);
-
-bool isSafeInFreeSpaceCollisionCheck(
-  const std::vector<std::pair<Pose, tier4_autoware_utils::Polygon2d>> & interpolated_ego,
-  const Twist & ego_current_twist, const std::vector<double> & check_duration,
-  const double prepare_duration, const PredictedObject & target_object,
-  const BehaviorPathPlannerParameters & common_parameters,
-  const double prepare_phase_ignore_target_speed_thresh, const double front_decel,
-  const double rear_decel, CollisionCheckDebug & debug);
 
 bool checkPathRelativeAngle(const PathWithLaneId & path, const double angle_threshold);
